+++ packages/terra-form-field/src/Field.jsx
@@ -202,7 +202,7 @@ const Field = (props) => {
     <div style={customStyles} {...customProps} className={fieldClasses}>
       {labelGroup}
       {content}
-      {isInvalid && error && <div aria-live="assertive" id={htmlFor ? `${htmlFor}-error` : undefined} className={cx('error-text')}>{error}</div>}
+      {isInvalid && error && <div id={htmlFor ? `${htmlFor}-error` : undefined} className={cx('error-text')}>{error}</div>}
       {help && <div id={htmlFor ? `${htmlFor}-help` : undefined} className={cx('help-text')}>{help}</div>}
     </div>
   );