+++ packages/terra-profile-image/tests/jest/ProfileImage.test.jsx
@@ -9,9 +9,21 @@ it('should render a Terra image with a placeholder prop', () => {
   expect(wrapper).toMatchSnapshot();
 });
 
+/* Profile Image with source and without ALt */
+it('should render a Terra image with a placeholder prop without alt prop', () => {
+  const wrapper = shallow(<ProfileImage src="profile.jpg" height="75" width="75" alt="" />);
+  expect(wrapper).toMatchSnapshot();
+});
+
 /* Profile Image source is not provided */
 it('should render an avatar image without Source prop', () => {
   const wrapper = shallow(<ProfileImage alt="Default" height="75" width="75" />);
+  expect(wrapper).toMatchSnapshot();
+});
+
+/* Profile Image without ALT */
+it('should render an avatar image without alt prop', () => {
+  const wrapper = shallow(<ProfileImage height="75" width="75" />);
   expect(wrapper).toMatchSnapshot();
 });