+++ packages/terra-form-field/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Changed
   * Updated jest snapshots for terra-icon changes.
   * Updated jest snapshots.
-  
+
 ## 4.20.5 - (July 5, 2022)
 
 * Changed