+++ packages/terra-form-field/CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 * Changed
-  * Set the hideRequired prop to private
-  
+  * Set the hideRequired prop to be private
+  * Set the isLabelHidden prop to be private
+
 ## 4.20.5 - (July 5, 2022)
 
 * Changed