+++ packages/terra-form-field/CHANGELOG.md
@@ -7,6 +7,9 @@
   * Set the isLabelHidden prop to be private
   * Set an asterisk denoting required field to be aria-hidden
 
+* Removed
+  * Removed aria-live attribute from error message.
+
 ## 4.20.5 - (July 5, 2022)
 
 * Changed